requirements.txt
msal==1.22.0
httpx==0.24.0
python-dotenv==1.0.0
google-auth==2.19.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.97.0

